+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Race condition due to unnecessary async calls causing status to be incorrect in few scenarios
 - Stress Tester would hang in some situations when some files returned with non-zero exit code
 - Interactive testcase status being overridden when saving new interactor secret during the run
+- Testcase runtime errors weren't detected when the interactor had failure causing it to not return an exit code
 
 ### Removed